Add flextable minibar to column

Add flextable minibar to column

add_score_minibar(
  flextable_df,
  score_col = "Weighted Score",
  risk_col = "Risk",
  column_index = NULL,
  color_text = FALSE,
  digits = 2
)

Arguments

flextable_df

a flextable object

score_col

The name of the column specifying the scores. Used to calculate

risk_col

The name of the column specifying the risks

column_index

Index of column to apply the minibar to. Traditionally, this should map to the index of score_col, but doesn't need to. If NULL, will map to the index of score_col.

color_text

Logical (T/F). If FALSE color the text in column_index black. Otherwise use risk colors (green, orange, red)

digits

number of digits to round scores to.

Details

a minibar will replace the entire column of whichever specified (via column_index). In other words, it doesn't append the minibar to a column. Specifying an index that doesn't correspond to the weighted scores, would overwrite that column.

digits must be specified, because the default behavior is rounding the column to one decimal