Check that package scores are valid

Check that individual scores are retained and category scores are numeric.

check_scores_valid(pkg_scores, json_path)

Arguments

pkg_scores

a named list containing the build up of score elements and overall category scores.

json_path

a JSON file path.

Details

Individual Score checks

The only checks performed on the individual scores are that they still exist when written out to json. This would only not be the case if an individual score was set to NULL, removing it from the list.

Category Score checks

covr scores are the only individual scores that allow for NA values. However when determining overall category scores, all values should be coerced to numeric, non-NA values (0 in the case of covr). Any potential errors or character values should be coerced to NA when passed through mean() in calc_overall_scores.

This function serves to confirm that all category scores are in fact numeric, non-NA values. The only way a category score could be anything other than a valid score or NA, is if mean() returned an error message or some other value. This scenario would likely cause issues before writing the score list out to json, so the additional is.numeric check is primarily a precautionary measure.