Render PDF summary of scorecards

Render PDF summary of scorecards

render_scorecard_summary(
  result_dirs,
  risk_breaks = c(0.3, 0.7),
  out_dir = NULL,
  snapshot = NULL,
  overwrite = TRUE
)

Arguments

result_dirs

A vector of output directories, each one produced by score_pkg(). external_scores are not supported.

risk_breaks

A numeric vector of length 2, with both numbers being between 0 and 1. These are used for the "breaks" when classifying scores into Risk categories. For example, for the default value of c(0.3, 0.7), for a given score: 0 <= score < 0.3 is "Low Risk", 0.3 <= score < 0.7 is "Medium Risk", and 0.7 <= score < 1 is "High Risk".

out_dir

Output directory for saving scorecard summary. If NULL, assumes all result_dirs point to the same output directory

snapshot

A report subtitle indicating the grouping of these packages, such as an MPN snapshot. See details.

overwrite

Logical (T/F). If TRUE, will overwrite an existing file path if it exists

Details

The argument snapshot, will appear in the document with language similar to 'MPN Snapshot <snapshot>'. In other words, snapshot should be a date, or some other grouping that makes sense grammatically according to the this example.