Wrap a vector of strings using specific characters

Wrap a vector of strings using specific characters

wrap_text(
  str,
  width = 23,
  wrap_sym = c("/", "_", "-"),
  strict = FALSE,
  indent = FALSE
)

Arguments

str

a string or vector of strings.

width

Positive integer giving the target line width (in number of characters).

wrap_sym

Characters to loop through and attempt to make new lines with. Note that the order you specify these values matters.

strict

logical (T/F). If FALSE, will soft wrap based on the characters specified via wrap_sym. If TRUE, will first soft wrap, and then enforce the specified width.

indent

logical (T/F). If TRUE, indent new lines via a markdown tab ('\t')

Details

stringr::str_wrap is not strict with the width for word characters, so a helper function was needed. Rather than being 100% strict at following the cutoff width, we attempt to split at white space and specific symbols commonly used in function names and file paths. If this splitting is not sufficient, we then perform a strict operation, whereby we cut the line off at exactly that width.